[#45] Refactor code. Implement reconnecting to hub
